Eliminated accidental React dependencies
