+++ packages/moxb/src/routing/index.ts
@@ -2,8 +2,6 @@ export {
     MyLocation,
     UpdateMethod,
     LocationManager,
-    useLocationManager,
-    LocationManagerProvider,
     TestLocation,
     QueryChange,
     BasicLocationManagerImpl,
@@ -33,8 +31,6 @@ export {
     defineNavRef,
     parseNavRef,
     LinkGenerator,
-    LinkGeneratorProvider,
-    useLinkGenerator,
     LinkGeneratorImpl,
 } from './navigation-references';
 
@@ -82,9 +78,7 @@ export {
 
 export { isTokenEmpty } from './tokens';
 
-export { TokenManager, TokenManagerProvider, useTokenManager } from './TokenManager';
+export { TokenManager } from './TokenManager';
 export { TokenManagerImpl, PathTokenMappingList } from './TokenManagerImpl';
 
 export { MultiStepProcess } from './MultiStepProcess';
-
-export { RoutingProvider } from './RoutingProvider';
