+++ packages/moxb/src/routing/location-manager/index.ts
@@ -3,8 +3,6 @@ export {
     LocationDescriptorObject,
     UpdateMethod,
     LocationManager,
-    useLocationManager,
-    LocationManagerProvider,
     QueryChange,
     TestLocation,
     SuccessCallback,
