+++ packages/moxb/src/routing/navigation-references/index.ts
@@ -1,4 +1,4 @@
 export { NavRef, NavRefCall } from './NavRef';
 export { NavRefProps, defineNavRef, parseNavRef } from './NavRefImpl';
-export { LinkGenerator, LinkGeneratorProvider, useLinkGenerator } from './LinkGenerator';
+export { LinkGenerator } from './LinkGenerator';
 export { LinkGeneratorImpl } from './LinkGeneratorImpl';
