+++ packages/react-html/src/globalContext.ts
