+++ packages/react-html/src/index.ts
@@ -17,3 +17,13 @@ export { NavRefRedirect } from './NavRefRedirect';
 export { NavLink, NavLinkProps, NavLinkParams } from './NavLink';
 export { BoundNavLink } from './BoundNavLink';
 export { CountingClock, CountingClockProps } from './CountingClock';
+export { createGlobalContext } from './globalContext';
+export {
+    LocationManagerProvider,
+    TokenManagerProvider,
+    LinkGeneratorProvider,
+    useLocationManager,
+    useTokenManager,
+    useLinkGenerator,
+    RoutingProvider,
+} from './routingProviders';
